Binary script now opens browser from command line
