+++ HTMLalyzer/templates/newFile.html
@@ -4,7 +4,7 @@
         <ul id="constantsList{0}">
         </ul>
     </div>
-    <div id="quantities{0}" style="display:flex;flex-direction:column;margin:0px">
+    <div class="quantitiesDiv " id="quantities{0}" style="margin:0px">
 		<table class="quantitiesTable" id="quantitiesTable{0}">
 			<thead>
 				<tr class="quantitiesTr">
